Server/requirements.txt
# fastapi==0.116.1
# uvicorn==0.33.0
# pydantic==2.10.6
# pydantic-core==2.27.2
# python-dotenv==1.0.1
# python-multipart==0.0.20
# requests==2.32.4
# httpx==0.28.1
# httpcore==1.0.9
# certifi==2025.7.14
# charset-normalizer==3.4.2
# idna==3.10
# urllib3==2.2.3
# sqlalchemy==2.0.43
# databases==0.9.0
# pydantic_settings==2.10.1
# asyncpg==0.30.0
# # pydantic[email]==2.0.0
# passlib==1.7.4
# google==3.0.0
# google_generativeai==0.8.5
# faiss_cpu==1.12.0
# bcrypt==4.3.0
fastapi==0.116.1
uvicorn==0.33.0
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.4
httpx==0.28.1
httpcore==1.0.9
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
urllib3==2.2.3
sqlalchemy==2.0.43
databases==0.9.0
asyncpg==0.30.0
pydantic[email]
passlib==1.7.4
bcrypt==4.2.0
google==3.0.0
google_generativeai==0.8.5
sentence-transformers==3.1.1
faiss_cpu==1.12.0